finish configuration store swap out
